requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django @ file:///tmp/build/80754af9/django_1625585912945/work
djangorestframework==3.13.1
idna==3.3
numpy==1.22.3
pandas==1.4.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.9
wincertstore==0.2
